Consolidate iis tests
