+++ src/Servers/IIS/IIS/test/Common.FunctionalTests/LogFileTests.cs
@@ -224,7 +224,7 @@ namespace Microsoft.AspNetCore.Server.IISIntegration.FunctionalTests
         public async Task CaptureLogsForOutOfProcessWhenProcessFailsToStart()
         {
             var deploymentParameters = Fixture.GetBaseDeploymentParameters(HostingModel.OutOfProcess);
-            deploymentParameters.TransformArguments((a, _) => $"{a} ConsoleWrite");
+            deploymentParameters.TransformArguments((a, _) => $"{a} ConsoleWriteSingle");
             var deploymentResult = await DeployAsync(deploymentParameters);
 
             var response = await deploymentResult.HttpClient.GetAsync("Test");
