--- src/Servers/IIS/IIS/test/testassets/StressTestWebSite/Program.cs
@@ -1,26 +1,0 @@
-// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Logging;
-
-namespace ANCMStressTestApp
-{
-    public class Program
-    {
-        public static void Main(string[] args)
-        {
-            var host = new WebHostBuilder()
-                .ConfigureLogging((_, factory) =>
-                {
-                    factory.AddConsole();
-                })
-                .UseKestrel()
-                .UseIISIntegration()
-                .UseStartup<Startup>()
-                .Build();
-
-            host.Run();
-        }
-    }
-}
