--- src/Servers/IIS/IIS/test/testassets/StressTestWebSite/StressTestWebSite.csproj
@@ -1,24 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <Import Project="..\..\..\..\build\testsite.props" />
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.0</TargetFramework>
-    <InProcessTestSite>true</InProcessTestSite>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Reference Include="Microsoft.AspNetCore.Server.IISIntegration" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Compile Include="..\shared\**\*.cs" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Reference Include="Microsoft.AspNetCore.ResponseCompression" />
-    <Reference Include="Microsoft.AspNetCore.Server.Kestrel" />
-    <Reference Include="Microsoft.Extensions.Logging.Console" />
-  </ItemGroup>
-
-</Project>
